src/main/java/com/vodafone/springboot/crud/domain/SequenceGeneratorServiceImpl.java
package com.vodafone.springboot.crud.domain;

import static org.springframework.data.mongodb.core.FindAndModifyOptions.options;
import static org.springframework.data.mongodb.core.query.Criteria.where;
import static org.springframework.data.mongodb.core.query.Query.query;

import com.vodafone.springboot.crud.domain.model.SensorEventModel;
import org.springframework.stereotype.Service;

import org.springframework.data.mongodb.core.MongoOperations;
import org.springframework.data.mongodb.core.query.Update;

import java.util.Objects;

import com.vodafone.springboot.crud.ports.SequenceGeneratorService;

@Service
public class SequenceGeneratorServiceImpl implements SequenceGeneratorService {

    private final MongoOperations mongoOperations;

    public SequenceGeneratorServiceImpl(MongoOperations mongoOperations) {
        this.mongoOperations = mongoOperations;
    }

    @Override
    public String generateSequence(String seqName) {
        SensorEventModel.DatabaseSequence counter = mongoOperations.findAndModify(query(where("_id").is(seqName)),
                new Update().inc("seq",1), options().returnNew(true).upsert(true),
                SensorEventModel.DatabaseSequence.class);

        return !Objects.isNull(counter) ? "sensor_12"+counter.getSeq() : "sensor_"+123;
    }
}
